Move PipelineTableMetaDataUtil
